Named state array.
For some reason, when this.state is an unnamed array, it gets converted from an array to a series of objects during setState(). I don't know why, but naming the array solved the problem. I'm guessing this is just a best practice kind of thing—don't leave unnamed shit in this.state.